Rebase with major improvements
The directory specfied in the compilation database for the current
translation unit. This made ccmock create its output file under the
wrong path if a relative path was used.

Previously ccmock had issues when generating mocks for functions
which had in no names for some of its arguments in its declaration.

Added system test using zlib-ng

Migrating README to reStructuredText

The relative paths written in the generated code comments were not
always correct. This is now fixed and additional the user is now able
to specify a base directory from were the relative paths shall be
calculated.

Changed test reports to XML format

Added support for globs as blacklist entries

As gmock and gtest might be static libaries on some systems,
the order in which libraries are passed to the linker is relevant
for the symbol lookup. This commit resolves an issue with a missing
symbol from the math library.